+++ test/test_llm_bedrock.py
@@ -2,7 +2,7 @@ import unittest
 import os
 
 from kegal import LlmBedrock
-import tests.llm.test_llm as test_llm
+from test import test_llm
 
 from pathlib import Path
 from dotenv import load_dotenv